Code Refactored Payment service
