+++ payment-service/src/main/java/com/epam/payment/PaymentApplication.java
@@ -1,13 +1,13 @@
-package com.epam.paymentservice;
+package com.epam.payment;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class PaymentserviceApplication {
+public class PaymentApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(PaymentserviceApplication.class, args);
+		SpringApplication.run(PaymentApplication.class, args);
 	}
 
 }
