+++ payment-service/src/main/java/com/epam/payment/controller/PaymentController.java
@@ -1,4 +1,4 @@
-package com.epam.paymentservice.controller;
+package com.epam.payment.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @RequestMapping("/v1/api")
-public interface PaymentServiceControllerInterface {
+public interface PaymentController {
 	
 	@PostMapping("/payments")
 	public String addPayment();
