+++ payment-service/src/main/java/com/epam/payment/repository/PaymentRepository.java
@@ -1,0 +1,5 @@
+package com.epam.payment.repository;
+
+public interface PaymentRepository {
+
+}
