+++ payment-service/src/main/java/com/epam/payment/service/PaymentClass.java
@@ -1,0 +1,5 @@
+package com.epam.payment.service;
+
+public class PaymentClass {
+
+}
