--- payment-service/src/main/java/com/epam/paymentservice/repository/PaymentServiceRepository.java
@@ -1,5 +1,0 @@
-package com.epam.paymentservice.repository;
-
-public interface PaymentServiceRepository {
-
-}
