--- payment-service/src/main/java/com/epam/paymentservice/service/PaymentServiceClass.java
@@ -1,5 +1,0 @@
-package com.epam.paymentservice.service;
-
-public class PaymentServiceClass {
-
-}
