+++ payment-service/src/test/java/com/epam/payment/PaymentApplicationTests.java
@@ -1,10 +1,10 @@
-package com.epam.paymentservice;
+package com.epam.payment;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PaymentserviceApplicationTests {
+class PaymentApplicationTests {
 
 	@Test
 	void contextLoads() {
